move identify_complex_polygons function as a method to ComplexPolygonsIdentifier class
